homework006_2
