+++ src/components/AnimalCard/AnimalCard copy.tsx
@@ -1,6 +1,8 @@
 import "./styles.css";
+import { AnimalCardProps } from './tyoes';
 
-function AnimalCard({ animalData}) {
+
+function AnimalCard({ animalData }: AnimalCardProps) {
   return (
     <div className="animal-card-wrapper">
       <h2>{animalData.name}</h2>
