+++ src/components/AnimalCard/tyoes.ts
@@ -1,0 +1,13 @@
+export interface AnimalCardProps {
+  animalData: any;
+  }
+
+export interface OldAnimalCardProps {
+  animalName: string;
+  species: string;
+  role: string;
+  animalImg: 'animalImg';
+  children: 'children';
+}
+
+
