+++ src/components/Input/tyoes.ts
@@ -1,0 +1,7 @@
+export interface InputProps {
+    label: string;
+    type: string;
+    name: string;
+    placeholder: string;
+    id: string;
+}
